tests: moved unit tests up Added parseSearchParams() test
