+++ test/barcode.js
@@ -1,7 +1,7 @@
 global.window = global;
 
 const assert = require("assert");
-const ripeCommons = require("../..");
+const ripeCommons = require("..");
 
 describe("encodeBarcode128B()", function() {
     it("should be able to create a code 128 value", () => {
